+++ apps/web/app/composables/useCategoryData/index.ts
@@ -1,0 +1,1 @@
+export * from './useCategoryData';
